only upgrade pip for new python installations

+++ extensions/notebook/src/jupyter/jupyterServerInstallation.ts
@@ -165,10 +165,12 @@ export class JupyterServerInstallation implements IJupyterServerInstallation {
 			let pythonExists = await utils.exists(this._pythonExecutable);
 			if (!pythonExists || forceInstall) {
 				await this.installPythonPackage(backgroundOperation, this._usingExistingPython, this._pythonInstallationPath, this.outputChannel);
+				// upgrade pip to make sure the right version is installed
+				if (!this._usingExistingPython) {
+					let cmd = `"${this._pythonExecutable}" -m pip install --upgrade pip`;
+					await this.executeBufferedCommand(cmd);
+				}
 			}
-			// upgrade pip to make sure the right version is installed
-			let cmd = `"${this._pythonExecutable}" -m pip install --upgrade pip`;
-			await this.executeBufferedCommand(cmd);
 			await this.upgradePythonPackages(forceInstall, packages);
 		} catch (err) {
 			this.outputChannel.appendLine(msgDependenciesInstallationFailed(utils.getErrorMessage(err)));
